requirements.txt
requests
beautifulsoup4
kafka-python
python-dotenv
aiohttp
